requirements.txt
SQLAlchemy==2.0.27
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
Flask==3.0.2
flask-cors==4.0.0
pymysql==1.1.0
gunicorn==21.2.0
flask-jwt-extended==4.6.0
pydantic==2.6.3
email-validator==2.1.1
celery==5.3.6
redis==5.0.1
b2sdk==1.25.0
sendgrid==6.11.0 